+++ packages/launchpad/src/lcdapi/utils.ts
@@ -1,7 +1,5 @@
+import { decodeBech32Pubkey, PubKey } from "@cosmjs/amino";
 import { Uint64 } from "@cosmjs/math";
-
-import { decodeBech32Pubkey } from "../pubkey";
-import { PubKey } from "../types";
 
 /**
  * Converts an integer expressed as number or string to a number.